Add InvokeGateway for abstraction of invoke method
The InvokeGateway interface was created to abstract the invoke method from '@tauri-apps/api/tauri'. Changes include addition of this new interface, implementation of this interface in 'TauriInvokeGateway.ts' and modification of 'GetVideoInfo.ts' to utilize this new structure. Unit tests were also adjusted to use a mock instance of the gateway.